3D Scanned Objects (.stl) of the African Diaspora
These files are 3D scans of objects related to the African Diaspora: a set of Ghanaian Yam Festival Twin figurines, a figurine of St. Benedict of Palermo from Brazil, and a carnival mask devil-figure from the Dominican Republic. These objects were scanned by the Vanderbilt Children's Hospital in January 2018 for 3D printing at Vanderbilt's Wond'ry for an exhibit there on the Slave Societies Digital Archive in Spring of 2018. The objects belong to Jane Landers and the scans may be used freely for educational purposes.